Refactor tests, setup.
