+++ .eslintrc.js
@@ -1,11 +1,14 @@
 module.exports = {
   root: true,
   extends: [
-    'digitalbazaar',
-    'digitalbazaar/jsdoc'
+    'eslint-config-digitalbazaar',
+    'eslint-config-digitalbazaar/jsdoc'
   ],
   env: {
-    node: true
+    node: true,
+    mocha: true
   },
-  ignorePatterns: ['dist/']
+  rules: {
+    'jsdoc/check-examples': 0
+  }
 };
