+++ lib/authorizeAccessToken.js
@@ -2,12 +2,11 @@
  * Copyright (c) 2021 Digital Bazaar, Inc. All rights reserved.
  */
 import * as authorization from 'auth-header';
-import bedrock from 'bedrock';
 import * as JWT from '@digitalbazaar/minimal-jwt';
-import {tokenizers} from 'bedrock-tokenizer';
+// import {tokenizers} from 'bedrock-tokenizer';
 import base64url from 'base64url-universal';
 
-const {util: {BedrockError}} = bedrock;
+const DEFAULT_ALLOWED_JWT_ALGS = new Set(['HS256', 'ES256']);
 
 /**
  * Authorizes an incoming request for a route (parses Authorization: header,
@@ -156,7 +155,7 @@ async function _verifyJwt({jwt, validIssuers}) {
   };
 }
 
-const ALLOWED_JWT_ALGS = new Set(['HS256', 'ES256']);
+
 
 async function _verifyFn({alg, kid, data, signature}) {
   if(!ALLOWED_JWT_ALGS.has(alg)) {
