+++ lib/noopLogger.js
@@ -1,4 +1,8 @@
 /*!
  * Copyright (c) 2021 Digital Bazaar, Inc. All rights reserved.
  */
-export {authorizeAccessToken} from './authorizeAccessToken.js';
+export default new Proxy({}, {
+  get() {
+    return () => {};
+  }
+});
