+++ package.json
@@ -18,42 +18,40 @@
   },
   "license": "BSD-3-Clause",
   "main": "lib/index.js",
+  "scripts": {
+    "lint": "eslint .",
+    "test": "npm run lint && npm run test-node",
+    "test-node": "cross-env NODE_ENV=test mocha -r esm --preserve-symlinks -t 10000 test/**/*.spec.js"
+  },
   "files": [
     "lib/*.js"
   ],
   "module": "lib/main.js",
   "dependencies": {
     "@digitalbazaar/minimal-jwt": "^1.0.1",
+    "assert-plus": "^1.0.0",
     "auth-header": "^1.0.0",
-    "base64url-universal": "^1.1.0",
-    "esm": "^3.2.25"
-  },
-  "peerDependencies": {
-    "bedrock": "^4.1.1",
-    "bedrock-https-agent": "^2.0.0",
-    "bedrock-mongodb": "^8.1.1",
-    "bedrock-tokenizer": "^1.1.0"
+    "base64url-universal": "^1.1.0"
   },
   "devDependencies": {
-    "eslint": "^7.20.0",
+    "chai": "^4.2.0",
+    "chai-http": "^4.3.0",
+    "cross-env": "^7.0.2",
+    "eslint": "^7.9.0",
     "eslint-config-digitalbazaar": "^2.6.1",
-    "eslint-plugin-jsdoc": "^32.0.1"
-  },
-  "nyc": {
-    "exclude": [
-      "tests"
-    ],
-    "reporter": [
-      "html",
-      "text-summary"
-    ]
+    "eslint-plugin-jsdoc": "^30.2.0",
+    "express": "^4.17.1",
+    "lru-cache": "^6.0.0",
+    "mocha": "^8.1.0"
   },
   "engines": {
     "node": ">=12"
   },
   "keywords": [
-    "Decentralized",
-    "Linked Data"
+    "oauth",
+    "oauth2",
+    "access",
+    "token"
   ],
   "directories": {
     "lib": "./lib"
