--- tests/.eslintrc.js
@@ -1,15 +1,0 @@
-module.exports = {
-  globals: {
-    should: true,
-    assertNoError: true
-  },
-  env: {
-    node: true,
-    mocha: true
-  },
-  extends: [
-    'digitalbazaar',
-    'digitalbazaar/jsdoc'
-  ],
-  ignorePatterns: ['node_modules/']
-};
