--- tests/package.json
@@ -1,41 +1,0 @@
-{
-  "name": "authorize-access-token-middleware-test",
-  "version": "0.0.1",
-  "private": true,
-  "scripts": {
-    "lint": "eslint .",
-    "test": "npm run lint && npm run test-node",
-    "test-node": "cross-env NODE_ENV=test node -r esm --preserve-symlinks test.js test",
-    "coverage": "cross-env NODE_ENV=test nyc --reporter=lcov --reporter=text-summary npm run test-node",
-    "coverage-ci": "cross-env NODE_ENV=test nyc --reporter=text-lcov npm run test-node > coverage.lcov",
-    "coverage-report": "nyc report"
-  },
-  "dependencies": {
-    "bedrock": "^4.1.1",
-    "bedrock-https-agent": "^2.0.0",
-    "bedrock-mongodb": "^8.1.1",
-    "bedrock-test": "^5.3.2",
-    "bedrock-tokenizer": "^1.1.0",
-    "esm": "^3.2.25"
-  },
-  "nyc": {
-    "excludeNodeModules": false,
-    "include": [
-      "node_modules/bedrock-vc-issuer/**"
-    ],
-    "exclude": [
-      "node_modules/bedrock-vc-issuer/node_modules/**"
-    ]
-  },
-  "devDependencies": {
-    "chai": "^4.3.0",
-    "chai-http": "^4.3.0",
-    "cross-env": "^7.0.3",
-    "eslint": "^7.20.0",
-    "eslint-config-digitalbazaar": "^2.6.1",
-    "express": "^4.17.1",
-    "lru-cache": "^6.0.0",
-    "mocha": "^8.3.0",
-    "nyc": "^15.1.0"
-  }
-}
