--- tests/test.config.js
@@ -1,21 +1,0 @@
-/*!
- * Copyright (c) 2021 Digital Bazaar, Inc. All rights reserved.
- */
-'use strict';
-
-const {config} = require('bedrock');
-const path = require('path');
-
-config.mocha.tests.push(path.join(__dirname, 'mocha'));
-
-// mongodb config
-config.mongodb.name = 'authorize_access_token_middleware_test';
-config.mongodb.host = 'localhost';
-config.mongodb.port = 27017;
-// drop all collections on initialization
-config.mongodb.dropCollections = {};
-config.mongodb.dropCollections.onInit = true;
-config.mongodb.dropCollections.collections = [];
-
-// HTTPS Agent
-config['https-agent'].rejectUnauthorized = false;
