--- tests/test.js
@@ -1,10 +1,0 @@
-/*!
- * Copyright (c) 2021 Digital Bazaar, Inc. All rights reserved.
- */
-const bedrock = require('bedrock');
-require('bedrock-mongodb');
-require('bedrock-tokenizer');
-require('bedrock-https-agent');
-
-require('bedrock-test');
-bedrock.start();
